Moved pattern types to new types package
