+++ packages/patterns/src/getPreviewText.ts
@@ -18,11 +18,10 @@
  * External dependencies
  */
 import { __ } from '@googleforcreators/i18n';
-
-/**
- * Internal dependencies
- */
-import { Pattern, PatternType } from './types';
+import {
+  Pattern,
+  PatternType,
+} from '@googleforcreators/types';
 
 function printRGB(r: number, g: number, b: number) {
   const hex = (v: number) => v.toString(16).padStart(2, '0');
