+++ packages/patterns/src/hasGradient.ts
@@ -15,9 +15,12 @@
  */
 
 /**
- * Internal dependencies
+ * External dependencies
  */
-import { Pattern, PatternType } from './types';
+import {
+  Pattern,
+  PatternType,
+} from '@googleforcreators/types';
 
 function hasGradient(pattern: Pattern) {
   return Boolean(pattern?.type) && PatternType.SOLID !== pattern.type;
